backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pandas>=2.1.0
numpy>=1.26.0
pydantic>=2.5.0
python-dotenv>=1.0.0
reportlab>=4.0.0
python-pptx>=0.6.23
matplotlib>=3.8.0
seaborn>=0.13.0
scipy>=1.11.0
Pillow>=10.1.0
aiofiles>=23.2.0
groq>=0.9.0
